move Diagram component to the Workbench folder
So it'll be placed in the place it's being used
